[IMP] l10n_cz: add VAT control statement report
Czech Republic requires a special tax declaration filled by the VAT payer. To implement it, a few fields need to be added to the account move, such as supply code and taxable supply date.

As of 2024, only one reduced rate is used, so this report already only implements two tax rates (as opposed to three tax rates used previously when the reduced and super reduced rates existed).
This commit also edits the tax rates and groups  to reflect this change.

task-3698642