[IMP] im_livechat, mail: allow to easily join/leave live chat channel
This PR provides a way for live chat users to efficiently join/leave a
live chat channel. Indeed, this operation is common to control the
number of chats an operator handle thus it should be as convenient as
possible. Currently, the operator must go to the live chat app and
click on the kanban button.

From now on, the operator will be able to either use the discuss
interface or the command palette to do so.

task-3640730