[FIX] web: adapt property field type selector style and z-index
The dropdown inside the property field popover style was broken due to
a dropdown refacor (https://github.com/odoo/odoo/pull/137691). The dropdown
opened by the popover was displayed behind the popover.

To fix this issue we added a supplementary div around each overlay items,
this isolates each overlay so their z-index do not interfere, their order
in the overlay-container is now sufficient to ensure a proper z-index,
whatever the amount of popovers, modals, dropdowns, etc...