[FIX] web: fixes optional dropdown issue
steps to reproduce:
--------------------------
- install project, sale and timesheet module.
- create a project without billable (remove checkmark on billable)
- after creating project create one task on created project.
- and then open that task.
- click on timesheets menu.
- in a list view click on three dots  in right corner of view.
- three dots still to appear even when no choices are available
  it shows  empty list ( blank box ).

issue :
-----------------
The optional dropdown (three dots) still  appear even when it's empty.

cause:
------------------
The function _shouldRenderOptionalColumnsDropdown() returns 'true' when no
fields are in the optional dropdown, and and it returns the number If one or
more fields are present in the optional dropdown.the present condition
accept both scenario that's a  reason the optional dropdown still appreing.

solution :
----------------------
a new condition is added in the main render function, _renderView().
This condition checks the type of value returned by
_shouldRenderOptionalColumnsDropdown(). If value is number and value greater
than 0, optional columns are added accordingly. This ensures that only when
optional fields are present, they are included, preventing the appearance of
unnecessary three dots in the list view.

task-3706742