[FIX] website_blog: remove illegitimate field translate
the teaser field was marked translate in PR #147358 which is illegitimate.
the 'translate' attribute was used only for website editor in the UI in the
'edit_translations' mode.

Instead of marking the field translate, this commit marks `lang`
context_dependent field with inverse method as translatable in the ir_qweb.py
Because these fields will behave like other translated fields