[FIX] portal: fix issues by the portal redesign
This commit is a follow up of Commit[1] that introduced a redesign of the
customer portal user interface.

While this redesign aimed to introduce a better layout and look of the
user interface, in also introduced some issues that needs to be addressed.
For a full list of the issues, please refer to the PR description.

> fix the sidebar layout 

> make the layout "dark mode proof"

> fix design inconsistencies

> improve the design of some views/elements that were not redesigned.



Commit[1]: e85d96e

task-3703251
follow up of 3046004