[FIX] web_editor: redirect protocol relative image src
Steps to reproduce:
- Create a db on version 16.0 (or before [this commit]).
- Upload an image on the website with an external protocol relative url
of an image that is CORS-protected (for example:
`//tinyjpg.com/images/social/website.jpg`).
- Save.
- Upgrade the db to this version.
- Click on the image.

-> Options that modify the image (like filter and quality) appear but
they should not as the image is CORS protected. Modifying one of those
options would lead the image to be displayed as a default one.

As the image was uploaded before [this commit], it is linked to an
attachment of type "url". Thanks to the previous commit, the system
manages to find this attachment. The problem is that `loadImageInfo()`
fills the dataset of the image with its `mimetype`, `originalId` and
`originalSrc` as it believes that the image is a local one. The bug is
fixed by marking all the attachments of type 'url' defined with a
protocol relative url as `redirected`. Thanks to it, the dataset of the
image is not filled and the options are not shown.

[this commit]: https://github.com/odoo/odoo/commit/943944dd249c15de870d6800d89e48d54a422e5a