[FIX] web_editor: popover left in DOM on test tear down
Slight changes!

Commit [1] extended the "link preview in Link Dialog" test by adding a
click to the link element, in order test the flow in which the user
edits the link via the button in the popover.

Such popover, when opening and closing, displays an animation that takes
about 150 ms to complete. If the test is run in a blazing fast
environment, the animation may not yet be complete when the test ends,
resulting in undesirable elements left in the body. In fact, this led
this test to fail in a nondeterministic fashion in some runbot
instances.

This commit makes sure such animation (CSS transitions) is complete
before moving on with the test.

[1]: https://github.com/odoo/odoo/commit/9db065e75d7b31febe4a42138308b47d0fa46508