[FIX] purchase_requisition: compute taxes for alternative PO creation
Steps to reproduce:

- Create a request for quotation for a product with set vendor Taxes
- Create an alternative purchase order from that RFQ with Copy Product

Current Behavior:

The purchase order line of the RFQ associated with the created
alternative did not compute the taxes.

Expected behavior:

The taxes should be computed in the same way as if the RFQ was created
manually and then linked to the alternatives.

Cause of the issue:

When creating a RFQ manually, the taxes of each purchase order line
are computed during the onchange method:
https://github.com/odoo/odoo/blob/64c13193ae0fe23eb1b426f5bbcd8531e9a40967/addons/purchase/models/purchase.py#L1221
https://github.com/odoo/odoo/blob/7ed0a773b6e0711b127f236ce665c35202d973bc/addons/purchase/models/purchase.py#L1055-L1061
when the product is set. However, creating an alternative will generate
each `purchase.order.line` of the RFQ using `Command.create` and will
therefore not trigger the onchange method computing the taxes.

Fix:

To KISS, we compute the taxes of each line just after the creation of
the record using the `_compute_tax_id` method.

opw-3750719