[FIX] website_mass_mailing: adjust rounded corners on snippet
Prior to this commit, the rounded corners  were missing on the subscribe
button. This is because this button is not the last element of the
`input-group`. A second button is placed next to it and is displayed
when the input is completed.

To fix this issue, we have to force the rounded corner of the subscribe
button.

task-3678772