[IMP] base: add lib ppf.datamatrix
Before this commit, we used `reportlab` to draw barcodes regardless the
kind of barcode we want. But while working with DataMatrix, `reportlab`
has some hardcoded limit making it unusable when the string is longer
than 144 bytes once it's encoded.

To have more freedom with DataMatrix, we decided to use another lib,
`ppf.datamatrix`. More information about the lib on its Github page:
https://github.com/adrianschlatter/ppf.datamatrix

Beside of that, to avoid making the `barcode` method a mess, a new
method for `ir.actions.report` is added: `datamatrix_svg`