[FIX] mrp_subcontracting: changing UoM causes unreserve issues
Before this commit: It was possible to purchase a subcontracted product
and on the receipt of the product, change the unit of measure o on the
stock move lines of the components when the quantities were already
reserved. This caused unreserve issue given that, all of a sudden,
instead of having 200g, we would have 200kg reserved and the change
would not be reflected on the quant.

After this commit: The view was modified to not allow changes of UoM if
there are reserved quantities, or if the state is done.

OPW-3742720

X-original-commit: c6c8e60