[FIX] analytic: limit applicability to specified (or active company)
Consider a an analytic plan in a multi-company environment.
This plan may have multiple applicability rules (account.analytic.applicability).
The default applicability of the plan and these applicability rules
are used to determine the applicability of the plan ('optional',
'mandatory' or 'unavailable').

For each applicability rule of an analytic plan a company can be specified.
Currently this company parameter is not considered when computing
the applicability of the plan.
This may render plans as 'unavailable' or 'mandatory' by applicability rules that
should only be used for other companies.
E.g. a plan may wrongly considered as mandatory for an invoice.
This can e.g. lead to an error when trying to confirm the invoice.
See 'Reproduce' below.

After this commit the company parameter rules is used to restrict the
applicability rules.

Reproduce:
  1. Enable analytic accounting in the settings.
  2. Ensure that there are 2 different companies: here C1 and C2
     Ensure the companies have a chart of accounts installed so that invoices can be posted.
  3. Create a new analytic plan AP1 with default applicabilty 'Optional'.
     (Accounting -> Configuration -> Analytic Plans)
  4. Create an applicability in the plan
       * Domain: Invoice
       * Applicability: mandatory
       * Company: A
  5. Create an analytic account for the plan that belongs to C1.
     This is the only account belonging to AP1.
  6. Create an invoice for C2 with a customer and a single line.
       * No analytic distribution can be set on this line.
         The analytic account from the previous step is not visible since it belongs to C1.
  7. Try to "Confirm" the invoice:
     -> Validation Error: "One or more lines require a 100% analytic distribution"

task-3770422