[FIX] various: be defensive with activity types, define master data
Mail: make "Todo" activity type master data users cannot remove as it is
required in various flows notably in fleet or used in various activity plans.
Also force its model to be False as it is used in various models and should
not suddenly be limited to a given model.

Task-3777606