[IMP] project: add 'project' breadcrumb in  project sharing
Currently, if the user wants to navigate back to the 'all projects'
view from the project sharing view, they have to manually add
'my/projects' to the browser's URL

In this commit, we have added the project to the first position
in the breadcrumb. As a result, the project will be add in the
breadcrumb, allowing the user to easily navigate back.

task-3368896