[FIX] osv: set id when value type NewId
Issue:
    - for a search domain (left, operator, right), when the operator is
    in HIERARCHY_FUNCS (child_of/parent_of), and right type is NewID, we
    get this error: "TypeError: 'NewId' object is not iterable"

Steps To Reproduce (in 17.0):
    - In a contact form open Studio and add a O2M field Customer (Task)
    - Create a new task in the O2M and set the Project to a project
    related to a SO.
    - Notice Traceback Error "TypeError: 'NewId' object is not iterable"

Explanation for this scenario:
    - In the search domain of _get_last_sol_of_customer , we parse
('order_partner_id', 'child_of', self.partner_id.commercial_partner_id.id),
 where the type of self.partner_id.commercial_partner_id.id is NewID
because we are currently editing the partner by adding a task.

    - Since the operator is in HIERARCHY_FUNCS the `parse` method calls
    the `to_ids` method which returns list(value) where value type is
    `NewId` thus the Traceback error.

I don't know a similar scenario in v 15.0 to add a test.
For example a situation where I have a record being edited
(its ID type is NewID), and there's a search operation with a domain
where the operator is child_of/parent_of and right is the record ID.

Solution:
    - check if the value is an instance of NewID and return value.origin