[FIX] account: enable move line ref multi edit
Currently, when trying to edit the reference field of multiple move lines at the same time, the changes are not applied.

Steps to reproduce:
-------------------
* Go to the **Acocunting** App
* Select **Accounting** > **Journal Items**
* Select multiple lines
* Edit the reference field
* Confirm
> Observation: the fields are not updated

Why the fix:
------------
The changes are not registered by the function `getChanges`.

https://github.com/odoo/odoo/blob/1cb0e73e3f0ca8c6ec49dd9217ded2f0b250339e/addons/web/static/src/views/relational_model.js#L702-L713

The first if condition is true which deletes `fieldName` from the changes. If we have a look at the function `isFieldOnly()`, we notice that is returns `True`.

https://github.com/odoo/odoo/blob/1cb0e73e3f0ca8c6ec49dd9217ded2f0b250339e/addons/web/static/src/views/relational_model.js#L766-L773

In our case `readonly = this.fields[fieldName]` as `"readonly" in activeField.modifiers` is false. `activeField.modifiers` is, in short, related to the xml, where `this.fields[fieldName]` is related to python. We can notice a discardance between python and the view.

In python, the field is readonly by default because it is a related fields, while in the view, the field was marked as `readonly=False`.

https://github.com/odoo/odoo/blob/1cb0e73e3f0ca8c6ec49dd9217ded2f0b250339e/addons/account/models/account_move_line.py#L65-L99

https://github.com/odoo/odoo/blob/1cb0e73e3f0ca8c6ec49dd9217ded2f0b250339e/addons/account/views/account_move_views.xml#L175

If want to mimic the behavior of version 16.0 and 17.0 we have to set the field as readonly in python. If not, the other solution would be to remove the `readOnly=False` from the xml view.

opw-3786963