[FIX] web: HOOT - let errors be handled by tests
Before this commit, errors would be entirely disabled by the test runner
before any other handler set on the window could receive it, like the
error service which listens for errors and rejections on the window to
display an error dialog.

This commit lets the error be dispatched to other listeners on the
window, and also allows the `toVerifyErrors` to also check the error's
cause for a string match (because Owl always wraps its errors in a
generic error without any relevant information).