[FIX] stock: missing move procurement group
The picking's `group_id` is related from the moves' one. And when a new
move is created, its group is got from the picking's one.  Before this
commit, this is done thanks to the `stock.move` `_default_group_id` but
it works only when the key `default_picking_id` is in the context, which
make it really fragile.

This commit replaces it by getting back the `group_id` from the picking
in the move's `create` in case the picking id is in the `vals`.
If not, the move should already get the procurement group from another
source (copied form another move for example.)
Do the same logic in the `write`.