[IMP] payment: improve allow unarchiving the archived tokens
Earlier, when a payment method (pm) or payment provider (pm) was deactivated the
tokens linked to it were also getting archived and they were prohibited from
being active again.

After this commit, the tokens can be unarchived but if the linked pm and the pp
if both are active.

task-3756355