[REF] web_editor, *: make SnippetsMenu an OWL Component
*: mass_mailing, mass_mailing_themes, test_website, website,
website_blog

This commit converts SnippetsMenu from a "legacy" widget to an OWL
Component. In doing so, it applies a few new principles that OWL
introduced mainly:

Do not propagate references to owl component instances anywhere.
Owl is responsible for components' lifecycle. All methods on owl
components should be considered private. Components can choose to
expose methods by passing them explicitly to whoever needs them.[1]

This means that everywhere in the code where a SnippetsMenu instance
was called (usually with `wysiwyg.snippetsMenu`), an alternative was
implemented:
- It was replaced by a prop were possible
- A bus is introduced which is passed to the SnippetsMenu as a prop. The
wysiwyg editor can then trigger events on that bus, and if it is async
await an onSuccess callback. **Note: no trigger should be done on that
but prior the original rendering of the SnippetsMenu**
- The code was moved directly on the SnippetsMenu so that a call to its
instance is no longer necessary.
- The mutex was moved to the wysiwyg, and is shared with the
SnippetsMenu, so that any needed call to it can do so without accessing
the (previously private) property.

However, though it is not common in OWL, this commit keeps the initial
rendering promise:
The wysiwyg awaits the original rendering with [2]. We kept this concept
by passing a Promise in a prop. This could be removed by adding another
callback.

This commit also takes advantage of the OWL rendering engine by moving
most of the SnippetsMenu UI into a static frontend template (previously
it was a backend view). The view is kept, to generate the list of
snippets and still generate the options' HTML.

The structure of the list of snippets is now the following:
```
<snippets>
    <category>
        <title>Base Structure
        <snippets>
            <t t-snippet=""...../>
        </snippets>
    </category>
</snippets>
```

Finally, tours are also adapted as the DOM is altered with those new
changes.

[1]: https://github.com/odoo/odoo/wiki/Javascript-coding-guidelines#general-guidelines
[2]: https://github.com/odoo/odoo/blob/c0b7312f47d22d036ff51e544d69e01ab83bdf73/addons/web_editor/static/src/js/wysiwyg/wysiwyg.js#L612

task-3323657