[FIX] mrp_account: Set bom_ids on account analytic
Link all the BoM to analytic account. Since the analytic distribution
on bom is company dependent we can't store it. The distribution is store
in a text property and we compute other field base on this when we need
it. The m2m between BoM and analytic account doesn't work because one is
computed and the other not but there is nothing store in it.

In order to fix it we would have to load all the BoM and browse their
property in order to find if the current account is in or not. Which is
bad for performance especialy when there is a lot of BoM. I suggest to
not do it for this reason compare to the gains it brings.