[REF] web_editor, *: use Theme/Design tab code across SnippetsMenus
*: mass_mailing, website

Commit [1] introduce the Theme tab on Website, and commit [2] introduced
the Design tab on MassMailing. Both share almost exactly the same code,
aside from the element to which the tab is attached. This has no
importance for the code, and it can be shared inside web_editor.

Therefore, this commit factorize the code inside web_editor.

[1]: https://github.com/odoo/odoo/commit/6fb4ed44a04008a81f860fa24a63d444d70f56b6
[2]: https://github.com/odoo/odoo/commit/a9e79f8bda8e18f44cfc823e5eb6b3b200de42b5

task-3323657