[REF] web_editor, mass_mailing: use props to hide remove format button
Prior to this commit, the "remove format" button was manually removed
from the toolbar, as in, its HTML element was removed.

This commit changes the approach to use the props of the toolbar.
Therefore, it adds a showRemoveFormat prop which is set to true
when the SnippetsMenu is folded, or if the "snippets" option is not
set.

This way, the HTMLElement of the toolbar is not modified by the
SnippetsMenu.

task-3323657