[FIX] hr: prevent error if record already exists
When the user creates a new "Employee" and sets a "Related User" in
"HR Settings" this traceback occurs.

Steps to reproduce:
- Install "hr" module
- Employees -> new Employees -> HR Settings -> select Related User
- Click on the Save icon

Traceback: ValueError
<class 'psycopg2.errors.UniqueViolation'>: "duplicate key value violates unique constraint "hr_employee_user_uniq"
DETAIL:  Key (user_id, company_id)=(9, 1) already exists.

Previously, it searched for employee records using the user ID and company ID as
the domain criteria, resulting in an error because records with the same user_id
and company_id already existed.
https://github.com/odoo/odoo/blob/9bab90fee9b58225b5f86a3ac65aa750b9d9b2ad/addons/hr/models/res_users.py#L301

This commit will fix the above error by adding a query that searches for
employees with specific conditions.

sentry - 5098073828