[IMP] stock: remove external ID for import export compatible
The quand id could change during the process. Each time a quant is empty
it will be deleted. Later when the inventory is filled again, it will be
created again with a new ID.

So the external Id is not suitable for import/export. It's better to
focus on other keys. Product, location, serial number/lot, package, owner.