[IMP] mail, website_slides: allow to configure mail template layout
For some mail template, we want to disable the mail header (displaying a quick
access button to the record) as it is not appropriated. As this header is
configured per model through an override of a mail_thread method
(_notify_get_recipients_groups) and we don't want to make it depends on the
template being sent, we add here a new encapsulating template without header
that can be forced per mail template (through the email_layout_xmlid field
added in odoo/odoo#114462). So, even if the model configure the header, it is
ignored if the "no header" template is forced.

While adding the new template, we factorize some parts of the encapsulating
templates (footer and "power by") to avoid repeating those parts in 3
templates.

Technical note: the added queries in the test test_karma_gain_multiple_course
(website_slides) is caused by the 2 added t-call in the mail template (for the
"powered by" and the "footer")

Task-2810103