[FIX] account_peppol: allow the download of UBL files w/o eas and endpoint
Currently, users are blocked from downloading the UBL file if the partner
or the company are missing Peppol EAS and Peppol Endpoint fields.
The checkbox is not visible and even if it was visible, the user would
see a UserError asking to fill those values.

Let users download these files but show a banner warning about missing
fields.