Fix test case
The original test in web_editor tested the "indentList" command.
Fun fact: this command is not used anywhere in the codebase, except for
the test suite...

The current test has as step function `keydownTab`, which, other than
indenting lists, indent non-list blocks as well.

For this reason I'm adapting the test.

PS: As of now, we don't have a "INDENT_LIST" command.