[FIX] mass_mailing: click discard button to reset changes
Before this commit :

In the mass_mailing, there existed an issue where clicking the discard
button failed to discard the changes made and discard button remaining visible.

After this commit :

This commit resolves the issue in the mass_mailing module on clicking
the discard button, properly discard the changes made and
the discard button is invisible.

task-3324903