[FIX] account: properly align activities summary in accounting dashboard
Version: 16.0

Issue
-----

The activities displayed by the misc. operations  view in the accounting
app take up the entire screen and span across the card if the activity summary
is too long.

Steps
-----

 - Open the Accounting App.
 - Create a new entry from 'Miscellaneous Operations'.
 - Add an activity with a long summary.
 - Go back the dashboard. The summary will try to span the entire width.

Cause
-----

The activities are displayed with a field tag in the kanban view. The field tag
refers to the activities js component as a widget. Due the use of a widget attribute,
the template generates a div with class: "o_field_widget". This class is defined
to have a css "display" attribe with a default value of "inline-block". Using "inline-block"
causes the activity row to take up the whole width.

Solution
--------

Override the display attribute of "o_field_widget" to have a value of "inline".

opw-3839992