[IMP] l10n_it: add support for Declaration of Intent
In Italy, an operator who exports more than 10% of all his revenues is considered a regular exporter.
Regular exporters can purchase without paying VAT within certain limits.

For this they have to do a Declaration of Intent with the authorities.
The regular exporter then gives the declaration information to his suppliers.
The supplier can then issue an invoice to the regular exported
without taxes by noting the declaration of intent.

This commit adds support for this use case.
It is now possible to specificy a declaration of intent on sales orders and invoices.
They are also stated on the respective PDFs and electronic invoices.
A dedicated tax and fiscal position was added to be applied in case a declaration of intent is used.

task-3794632