[FIX] mrp: prevent valueerror when receiving unexpected values
This error arises when the values expected in "move_vals" are not received as
anticipated, leading to a situation where the system encounters an unexpected
value.

Traceback:ValueError
not enough values to unpack (expected 3, got 2)

https://github.com/odoo/odoo/blob/e2c6b088f0d774451446986339cf6dcf05db665c/addons/mrp/models/mrp_production.py#L865
Here, vals['move_raw_ids'] contains 2 values, so for 'field_values' there is no
value. So when the user tries to save, It will lead to the traceback.

similar issues PRs :
https://github.com/odoo/odoo/commit/c366efa6752995d321f59b395814bf28fb61afc2

sentry - 5128761097