[MOV] core, web_editor: move diff utils to core
These utils are currently used in dependencies of `web_editor`, but they
don't make use of any of the `web_editor` features.

A future commit will make use of these utils also in `mail`, which is
not dependent on `web_editor`.

Candidate modules for the move are: `mail`, `bus`, `web`, or `base`. As
these utils are not linked to any particular module, `base` seems to be
the most apprioriate of them.