[MOV] web_editor, web: move History Dialog to web
This dialog is currently used in dependencies of `web_editor`, but it
doesn't make use of any of the `web_editor` features.

A future commit will make use of these utils also in `mail`, which is
not dependent on `web_editor`.

Candidate modules for the move are: `mail`, `bus`, or `web`. As this
dialog is not linked to any particular module, `web` seems to be the
most apprioriate of them.