[IMP] web_editor: improve UX of history dialog
Currently, the history dialog has a shared scroll view for the list of
revisions and the diff view. When one of the columns has more content
than the other, the shorter column becomes invisible when scrolling down
the view, which can be annoying.

Example: When there are a lot of revisions and the user picks one at the
bottom of the scroll view, the user has to scroll back to the top to read
the diff of that revision.

To improve the usability of the history dialog, we will ensure that the
two columns have their own scroll view. People will then be able to
scroll independently in the revision list and the diff view. This commit
will also remove the statement blocking the user interface and creating
a flicker.

task-3853374