[REF] web_editor: move SnippetOptionWidget into legacy file
This will allow both the OWL Option system and the Legacy system to work
alongside each other. Making it easier to convert step by step.

The original file will be the one used for the conversion. This will
allow any fix done in the master branch to hit the OWL file when rebased
since this is the file that will be kept at the end of the conversion.

task-3850413

[REF] website: split snippet options to legacy file

task-3850413